fix reopened SAK-8737
Cancel button will reset all selections and typed text into initial (defaults).

Customized error for empty subject will be checked both when "Send" button is clicked and when "Attach a file"/"Attach another file" is clicked

...

More friendly error message is programmed but not applied since changing a message need substantial programming with little effect and also make it hard to internationalize the messages when hardcorded in the java source.

It will be addressed in the discussion of future version (possibly 2.5)

git-svn-id: https://source.sakaiproject.org/svn/mailtool/branches/sakai_2-4-x@32251 66ffb92e-73f9-0310-93c1-f5514f145a0a